% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest.R
\name{aml_gbm}
\alias{aml_gbm}
\title{AML Gradient Boosted Machine (GBM)}
\usage{
aml_gbm(data, response, learning_rate = 0.1, n_trees = 10, m = NULL,
  evaluation_criterion = sum_of_squares, min_obs = 5, max_depth = 8)
}
\arguments{
\item{data}{Input data.frame of dimension n x p for training the GBM}

\item{response}{Response vector of size nx1 corresponding to the training
data}

\item{learning_rate}{Shrinkage factor used to dictate learning speed, 
defaults to .1}

\item{n_trees}{Number of trees to train, defaults to 1000}

\item{m}{Number of columns to randomly use at each splitting iteration, 
defaults to all columns}

\item{evaluation_criterion}{Function that calculates error criterion for
fitting, defaults to sum of squares}

\item{min_obs}{Minimum observations allowed to end up in a single node,
defaults to 5}

\item{max_depth}{Maximum number of successive splits allowed to happen
in the tree, defaults to 8}
}
\value{
Results trained list of class aml_random_forest filled with random forest trees
}
\description{
Trains an ensemble of trees via gradient boosting
}

