% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_means.R
\name{aml_k_means}
\alias{aml_k_means}
\title{AML k-Means}
\usage{
aml_k_means(data, k, maximum_iterations = 1e+06)
}
\arguments{
\item{data}{Input n x p sized data.frame of numeric data}

\item{k}{Number of clusters to fit}

\item{maximum_iterations}{Maximum number of iterations to allow while trying
to converge}
}
\value{
Results object of class aml_k_means
}
\description{
Calculates k means predictions for given data. It is assumed that all
given columns will be used in calculation.
}

