% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_network.R
\name{aml_neural_network}
\alias{aml_neural_network}
\title{AML neural network}
\usage{
aml_neural_network(sizes, learning_rate, data, response, epochs,
  verbose = FALSE)
}
\arguments{
\item{sizes}{Vector of layer sizes}

\item{learning_rate}{Learning rate for gradient descent}

\item{data}{Input data.frame for training network, 
dimension nxp where p is sizes[1]}

\item{response}{Response vector of size nx1 corresponding to the training
data}

\item{epochs}{Number of epochs to use in network training}

\item{verbose}{Logical option for status printing}
}
\value{
Trained neural network object of class aml_neural_network
}
\description{
Trains a simple back-propagation neural network.
}

