% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest.R
\name{aml_random_forest}
\alias{aml_random_forest}
\title{AML Random Forest}
\usage{
aml_random_forest()
}
\value{
Results object of class aml_random_forest
}
\description{
Calculates predictions via random forest algorithm
}

