% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_network.R
\name{predict.aml_neural_network}
\alias{predict.aml_neural_network}
\title{AML neural network predict method}
\usage{
\method{predict}{aml_neural_network}(network, data)
}
\arguments{
\item{network}{Results network object returned by aml_neural_network function}

\item{data}{Data of correct size to be fed into the neural network}
}
\value{
Vector of predictions corresponding to the data rows
}
\description{
Returns predictions for a neural network when given a data set.
}

