% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest.R
\name{predict.aml_random_forest}
\alias{predict.aml_random_forest}
\title{AML random forest predict method}
\usage{
\method{predict}{aml_random_forest}(forest, data)
}
\arguments{
\item{forest}{Forest of trees trained by `aml_random_forest`}

\item{data}{Data.frame row of size `1 x p` for prediction}
}
\value{
Prediction value
}
\description{
Returns prediction for a randomm forest when given a row of data.
}

