% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest.R
\name{predict.aml_tree}
\alias{predict.aml_tree}
\title{AML single tree predict method}
\usage{
\method{predict}{aml_tree}(tree, data)
}
\arguments{
\item{tree}{Single fitted tree returned by the `create_tree` function}

\item{data}{Data.frame row of size `1 x p` for prediction}
}
\value{
Prediction value
}
\description{
Returns prediction for a single tree when given a row of data.
}

