% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{predict_all}
\alias{predict_all}
\title{AML predict method for more than one observation, works for GBM and RF}
\usage{
predict_all(ensemble, data, n_trees = NULL)
}
\arguments{
\item{ensemble}{Ensemble of trees trained by `aml_gbm`}

\item{data}{Data.frame row of size `n x p` for prediction}

\item{n_trees}{Number of trees used in prediction, uses all trees by default}
}
\value{
Vector of predictions
}
\description{
Returns prediction for a GBM/RF when given data frame
}

