% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{aml_random_forest}
\alias{aml_random_forest}
\title{AML Random Forest}
\usage{
aml_random_forest(data, response, b, m = NULL,
  evaluation_criterion = sum_of_squares, min_obs = 5, max_depth = 8,
  verbose = FALSE)
}
\arguments{
\item{data}{Input data.frame of dimension n x p for training the random forest}

\item{response}{Response vector of size nx1 corresponding to the training
data}

\item{b}{Number of bootstrap iterations to perform (trees to build)}

\item{m}{Number of columns to randomly use at each splitting iteration, 
defaults to all columns}

\item{evaluation_criterion}{Function that calculates error criterion for
fitting, defaults to sum of squares}

\item{min_obs}{Minimum observations allowed to end up in a single node,
defaults to 5}

\item{max_depth}{Maximum number of successive splits allowed to happen
in the tree, defaults to 8}

\item{verbose}{Flag to display training updates in the console}
}
\value{
Results trained list of class aml_random_forest filled with random forest trees
}
\description{
Trains an ensemble of trees via random forest
}

