% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{predict.aml_random_forest}
\alias{predict.aml_random_forest}
\title{AML random forest predict method}
\usage{
\method{predict}{aml_random_forest}(forest, data, n_trees = NULL)
}
\arguments{
\item{forest}{Forest of trees trained by `aml_random_forest`}

\item{data}{Data.frame row of size `1 x p` for prediction}

\item{n_trees}{Number of trees used in prediction, uses all trees by default}
}
\value{
Prediction value
}
\description{
Returns prediction for a randomm forest when given a row of data.
}

